pyproject.toml
[tool.poetry]
name = "rag-pipeline"
version = "0.1.0"
description = "Retrieval-Augmented Generation pipeline with document ingestion and querying"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.10"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pinecone-client = "^3.0.0"
openai = "^1.10.0"
google-generativeai = "^0.3.2"
tiktoken = "^0.5.2"
pymupdf = "^1.23.8"
python-docx = "^1.1.0"
pdfminer-six = "^20221105"
python-magic = "^0.4.27"
slowapi = "^0.1.9"
redis = "^5.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}
httpx = "^0.26.0"
tenacity = "^8.2.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
ruff = "^0.1.14"
black = "^24.1.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing --cov-report=html --cov-fail-under=85"

